+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 # Version 19
 
 ### [Current changes]
+- **[Add]** Custom error msg feature
 - **[Change]** Support for the new Authentication System 2.0 has been added.
 - **[Change]** Improved reconnect time of node by 10x-30x once it detects a live server.
 - **[Change]** Removed unnecessary codes + APIs.